Add magnetic field control point and offset for iso control points (#68)
* Add magnetic field control point and offset for iso

This is intended to be used for setting probe and flux loop control
targets into FRESCO.

* Fix issues introduced by merge

* Add test for all control point types and fix issues

The new control points now appear to be working properly.

* Add sigma_RpZ factors to field_cps

Thanks to Brendan for pointing out the missing COCOs factor

* Cleanup suggested during code review

Thanks to Brendan for the detailed review

* Revert change to test tolerance

The test passes on my laptop but the GitHub runner seems to have
larger error.

* Revert more changes to test tolerance

There are actually zero comparisons in the flux and field cps so
atol still needs to be used.

* Remove IsoRefControlPoints export

---------

Co-authored-by: Brendan Lyons <lyonsbc@fusion.gat.com>