Apply averages to coil and magnetic data
This uses the new averaging functions in an IMAS repo PR. The data
is averaged over the same time window that is used in EFIT, which
could be made into input variables if there is interest in more
easily varying it. These averages are currently just over independent
windows, which differs from EFIT and different options may be added
in the future.